move get_caller_module_name and is_call_from_module to generics
